# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api structure",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/manage/branch/{id}": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get specific Branch data",
                "description": "Branch show",
                "operationId": "showBranch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID to fetch the targeted campaigns.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Branch"
                ],
                "summary": "Branch update",
                "description": "Update specific branch",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID to fetch the targeted campaigns.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Bukhara"
                                    },
                                    "location": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "address"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Branch"
                ],
                "summary": "Branch delete",
                "description": "Delete specific branch",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID to fetch the targeted campaigns.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/manage/schedule": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Get all schedules data",
                "description": "Schedule index",
                "operationId": "indexSchedule",
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Set new schedule",
                "description": "Schedule store",
                "operationId": "storeSchedule",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "group_id",
                                    "weekday_id",
                                    "session_id",
                                    "room_id"
                                ],
                                "properties": {
                                    "group_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "weekday_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "session_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "numeric",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/manage/schedule/{id}": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Get specific schedule data",
                "description": "Schedule show",
                "operationId": "showSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID to fetch the targeted campaigns.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Update specific schedule",
                "description": "Schedule update",
                "operationId": "updateSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID to fetch the targeted campaigns.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "group_id",
                                    "weekday_id",
                                    "session_id",
                                    "room_id"
                                ],
                                "properties": {
                                    "group_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "weekday_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "session_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "numeric",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Schedule delete",
                "description": "Delete specific Schedule",
                "operationId": "destroySchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID to fetch the targeted campaigns.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/manage/schedule/by-filter": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Get schedules by-filter",
                "description": "Get schedules by-filter",
                "operationId": "getSchedulesByFilterSchedule",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "group_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "weekday_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "session_id": {
                                        "type": "numeric",
                                        "example": 1
                                    },
                                    "room_id": {
                                        "type": "numeric",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}